requirements.txt
numpy==1.19.4
pandas==1.1.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
